+++ src/main/java/app/repository/SkillRepository.java
@@ -2,5 +2,5 @@ package app.repository;
 
 import app.model.Skill;
 
-public interface SkillRepository extends GenericRepository<Skill, Long> {
+public interface SkillRepository extends GenericRepository<Skill, Integer> {
 }
